apps/api/requirements.txt
fastapi==0.104.1
httpx==0.25.2
aiohttp==3.9.1
psutil==5.9.6
uvicorn[standard]==0.24.0
pydantic==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
passlib[bcrypt]==1.7.4
redis==5.0.1
beautifulsoup4==4.12.2
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-logging==0.43b0

